requirements.txt
gradio
requests
flask
huggingface-hub>=0.26.0
python-dotenv
